+++ adapters/gorabbitstream/integration_test.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/faustbrian/go-rabbitmq-streams"
-	rabbitmqadapter "github.com/faustbrian/go-rabbitmq-streams/rabbitmq"
+	rabbitmqadapter "github.com/faustbrian/go-rabbitmq-streams/adapters/rabbitmq"
 	"github.com/faustbrian/go-transactional-outbox"
 	"github.com/faustbrian/go-transactional-outbox/adapters/gorabbitstream"
 	"github.com/rabbitmq/rabbitmq-stream-go-client/pkg/stream"